Use JPEG compression for stored GeoTIFFs.
JPEG compression gives drastically smaller files (anywhere from 1/4 to
1/10 the size of the deflated ones).  Mapnik handles the JPEG-compressed
GeoTIFFs just fine; the rendered images look the same as the ones from
deflate-compressed GeoTIFFs.  ImageMagick does not do as well, so the
prep_toposm_data script has to make a uncompressed version of the
hillshade and colormap files to be used to derive the hypsorelief file.

Also, since the colormap files are not used in rendering, I don't bother
saving a JPEG-compressed copy of those (and the other copies are deleted
after they're no longer needed).